bot-app/pyproject.toml
[tool.poetry]
name = "botapp"
version = "0.1.0"
description = "telegram bot for booking a photo shoot"
authors = ["Valeriya Sidorova <va_sidorova@inbox.ru>", "optikrus <optikrus61@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
aiogram = "^3.0.0"
pydantic-settings = "^2.0.3"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
ruff = "^0.1.6"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
aerich = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
